refactor : create 리팩토링
