+++ src/main/java/com/hangout/hangout/domain/comment/service/CommentService.java
@@ -1,11 +1,13 @@
-package com.hangout.hangout.Commnet.service;
+package com.hangout.hangout.domain.comment.service;
 
-import com.hangout.hangout.Commnet.domain.repository.CommentRepository;
-import com.hangout.hangout.Commnet.dto.CommentSaveRequestDto;
-import com.hangout.hangout.Commnet.dto.CommentUpdateRequestDto;
+import com.hangout.hangout.domain.comment.domain.repository.CommentRepository;
+import com.hangout.hangout.domain.comment.dto.CommentSaveRequestDto;
+import com.hangout.hangout.domain.comment.dto.CommentUpdateRequestDto;
 import com.hangout.hangout.domain.comment.entity.Comment;
+import com.hangout.hangout.domain.post.entity.Post;
+import com.hangout.hangout.domain.user.entity.User;
+import com.hangout.hangout.global.common.domain.entity.Status;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,8 +17,8 @@ public class CommentService {
     private final CommentRepository commentRepository;
 
     @Transactional
-    public void saveComment(CommentSaveRequestDto comment){
-        Comment comment1 = comment.toEntity();
+    public void saveComment(CommentSaveRequestDto comment, Post post, Status status, User user1){
+        Comment comment1 = comment.toEntity(comment, post,status,user1);
         commentRepository.save(comment1);
     }
