+++ src/main/java/com/hangout/hangout/global/annotation/LoginMember.java
@@ -1,0 +1,9 @@
+package com.hangout.hangout.global.annotation;
+
+import java.lang.annotation.*;
+
+@Target({ElementType.PARAMETER})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface LoginMember {
+}
